create: resolve TPM_RC_ATTRIBUTES for the sealed data
According to the spec, the sealed data cannot be an empty buffer if
both sign and decrypt are clear.

Certain dTPM chip such as Infineon TPM SLB 9665 doesn't strictly follow
the spec and allows to set the an empty buffer for the sealed data but
we should not reply on it.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>