init.cryptfs,luks-setup.sh: Improve startup reliability
Try clearing the tpm lockout before giving up unsealing the disk key
for the init.cryptfs.

In luks-setup.sh make use of /dev instead of /tmp because /tmp may not
be mounted at the time that luks-setup is called in mapping mode.
Also make use of tcti-probe when using mapping mode with the resourcemgr.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>